+++ src/oarepo_model/presets/drafts/files/draft_file_mapping.py
