+++ src/oarepo_model/presets/drafts/files/draft_with_media_files.py
