+++ src/oarepo_model/presets/drafts/files/record_with_media_files.py
